+++ Sources/FilterTag.swift
@@ -4,7 +4,7 @@ class FilterNode : NodeType {
   let token: Token?
 
   class func parse(_ parser: TokenParser, token: Token) throws -> NodeType {
-    let bits = token.components()
+    let bits = token.components
 
     guard bits.count == 2 else {
       throw TemplateSyntaxError("'filter' tag takes one argument, the filter expression")